2.0.1
This is my first time that I doing this This is also my first audio that I trying to make it what I can, but this is an alpha version, however. to make it as an exe file. it's difficult to me for me to make it so I'm splitting in a parts. and you need to download it all of one. I've been trying to follow you to tutorials even chat GPT. but all of them is miss failing to understanding. So if you want better helping me you can always contact me.